refactor and add tests, make a few fields public
